requirements.txt
fastapi
uvicorn
pydantic
jinja2
neo4j
python-dotenv
faiss-cpu
sentence-transformers
spacy
openai